src/requirements.txt
streamlit>=1.29  #frontend framework for web apps
supabase>=2.0.2     #supabse client for database operations
fastapi>=0.104.1    #backend api framework
uvicorn>=0.24.0     #ASGI server for  FATAPI
python-dotenv>=1.0.0    #Environment variable management
PyPDF2
httpx
